big bang.
This is the first main commit for this tool, which adds a view and less
command.

The view command will format a tab/csv file. This is suitable for
printing on the console or sending through less.

The less command is a more interactive viewer. You can scroll through
lines, and search for strings.